Switched content from about page to officers page
